+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Modify Prometheus exporter to translate non-monotonic Sums into Gauges
   ([#3306](https://github.com/open-telemetry/opentelemetry-python/pull/3306))
+- Fix error when no LoggerProvider configured for LoggingHandler
+  ([#3372](https://github.com/open-telemetry/opentelemetry-python/pull/3372))
 
 
 ## Version 1.19.0/0.40b0 (2023-07-13)